+++ data_tools/embeddings_db.py
@@ -7,8 +7,8 @@ from typing import Optional
 import numpy as np
 from tqdm import tqdm
 
-from data_models.embedders.embedder import Embedder
-from data_models.textsnippet import TextSnippet
+from data_tools.embedders.embedder import Embedder
+from data_tools.textsnippet import TextSnippet
 from experiments.utils.paths import default_embeddings_path
