+++ experiments/bjoern/calc_msg_embeddings.py
@@ -1,9 +1,9 @@
 from typing import List
 
-from data_models.embedders.embedder_transformer import EmbedderTransformer
-from data_models.embeddings_db import EmbeddingsDB
-from data_models.message_db import MessageDB
-from data_models.textsnippet import TextSnippet
+from data_tools.embedders.embedder_transformer import EmbedderTransformer
+from data_tools.embeddings_db import EmbeddingsDB
+from data_tools.message_db import MessageDB
+from data_tools.textsnippet import TextSnippet
 from experiments.utils.paths import default_embeddings_path
 from experiments.utils.paths import default_jsonl_path
