+++ experiments/bjoern/exp_count_practice_logs.py
@@ -1,5 +1,5 @@
-from data_models.categories import DhOCategory
-from data_models.message_db import MessageDB
+from data_tools.dho_categories import DhOCategory
+from data_tools.message_db import MessageDB
 from experiments.utils.paths import default_jsonl_path
