+++ experiments/bjoern/exp_model_comparison.py
@@ -8,8 +8,8 @@ import pandas as pd
 import plotly.express as px
 from sklearn.decomposition import PCA
 
-from data_models.categories import DhOCategory
-from data_models.message_db import MessageDB
+from data_tools.dho_categories import DhOCategory
+from data_tools.message_db import MessageDB
 from experiments.utils.paths import default_embeddings_path
 from experiments.utils.paths import default_jsonl_path
