+++ experiments/experiment_template.py
@@ -1,6 +1,6 @@
 import plotly.express as px
 
-from data_models.message_db import MessageDB
+from data_tools.message_db import MessageDB
 from experiments.utils.paths import default_jsonl_path
